LPS-55527 Move S3Store class to its module
