+++ portal-impl/src/com/liferay/portlet/documentlibrary/store/StoreFactory.java
@@ -179,9 +179,6 @@ public class StoreFactory {
 		},
 		new String[] {
 			"com.liferay.documentlibrary.util.JCRHook", JCRStore.class.getName()
-		},
-		new String[] {
-			"com.liferay.documentlibrary.util.S3Hook", S3Store.class.getName()
 		}
 	};
