+++ portal-impl/src/portal.properties
@@ -8988,7 +8988,6 @@
     #dl.store.impl=com.liferay.portlet.documentlibrary.store.DBStore
     dl.store.impl=com.liferay.portlet.documentlibrary.store.FileSystemStore
     #dl.store.impl=com.liferay.portlet.documentlibrary.store.JCRStore
-    #dl.store.impl=com.liferay.portlet.documentlibrary.store.S3Store
 
     #
     # AdvancedFileSystemStore
